Cambio de Int a Double para la propiedad Objetivo

+++ PMMX.Modelo/Vistas/CRRView.cs
@@ -9,7 +9,7 @@ namespace PMMX.Modelo.Vistas
     public class CRRView
     {
         public Double CRR_Total { get; set; }
-        public int Objetivo { get; set; }
+        public Double Objetivo { get; set; }
         public DateTime Fecha { get; set; }
     }
 }
